feat(defaultColumns): add simple regex support and move to service
